pyproject.toml
[project]
name = "langtrader-v0-2-0"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ccxt>=4.5.22",
    "langchain>=1.1.2",
    "langchain-anthropic>=1.2.0",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.4",
    "loguru>=0.7.3",
    "pandas>=2.3.3",
    "pandas-ta>=0.4.71b0",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "sqlmodel>=0.0.27",
    "websockets>=15.0.1",
]
